SchlierenReconstructions.jl

Documentation for SchlierenReconstructions.

SchlierenReconstructions.LSQR

SchlierenReconstructions.LeastSquaresProblem

SchlierenReconstructions.MatrixOperator

SchlierenReconstructions.solve

SchlierenReconstructions.LSQR — Type

LSQR <: AbstractSolver

Flag object for the multiple dipatch for function solve. solve(prob, LSQR()) means that prob will be solved with the LSQR method (Paige & Saunders (1982) "LSQR: An Algorithm for Sparse Linear Equations and Sparse Least Squares," ACM Transactions on Mathematical Software 8(1) 43-71. doi: 10.1145/355984.355989)

関数solveの多重ディスパッチ用のフラグオブジェクト． solve(prob, LSQR())は，probがLSQR法で解かれることを意味する．

SchlierenReconstructions.LeastSquaresProblem — Type

LeastSquaresProblem(operator, bvec)

Storing information on a least squares problem $\mathrm{argmin}_\boldsymbol{x} \left\Vert \boldsymbol{b} - A\boldsymbol{x} \right\Vert_2^2$

operator::T where T<:AbstractOperator corresponding to $A \in \mathbb{R}^{m\times n}$

bvec::T where T<:Vector corresponding to $\boldsymbol{b} \in \mathbb{R}^m$

最小二乗問題 $\mathrm{argmin}_\boldsymbol{x} \left\Vert \boldsymbol{b} - A\boldsymbol{x} \right\Vert_2^2$ についての情報を保存する構造体．

operator::T where T<:AbstractOperator は$A \in \mathbb{R}^{m\times n}$に対応

bvec::T where T<:Vector は$\boldsymbol{b} \in \mathbb{R}^m$に対応

SchlierenReconstructions.MatrixOperator — Type

MatrixOperator(A, dim_range, dim_domain) <: AbstractOperator

Flag object for the multiple dipatch of op! and op_adj!.

op!(y, x, MatrixOperator(A)) means that y = A*x,

A <: Matrix corresponds to $A \in \mathbb{R}^{m\times n}$

dim_range corresponds to $m$ (dimension of y)

dim_domain corresponds to $n$ (dimension of x)

関数op!, op_adj!の多重ディスパッチのためのフラグオブジェクト． op!(y, x, MatrixOperator(A))は，y = A*xの計算をすることを意味する．

A <: Matrix, $A \in \mathbb{R}^{m\times n}$

dim_range, $m$ (yの次元)

dim_domain, $n$ (xの次元)

SchlierenReconstructions.solve — Method

solve(prob::LeastSquaresProblem, solver::LSQR; ...)

Solve a least squares problem prob with LSQR method. (Paige & Saunders (1982) "LSQR: An Algorithm for Sparse Linear Equations and Sparse Least Squares," ACM Transactions on Mathematical Software 8(1) 43-71. doi: 10.1145/355984.355989)

Note: Implemented based on scipy.sparse.linalg.lsqr https://github.com/scipy/scipy/blob/v1.9.0/scipy/sparse/linalg/_isolve/lsqr.py#L96-L586

n_step_max, the maximum number of iteration steps

xvec_init, initial guess of the solution x (must be the same size as x)